search_agent/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
langgraph
langgraph-checkpoint-sqlite
python-dotenv
requests
beautifulsoup4